Implement header state to track which page user is on
