+++ src/components/Header.jsx
@@ -1,11 +1,42 @@
 import React, { Component } from 'react';
+import classNames from 'classnames';
 import { goToAnchor, removeHash } from 'react-scrollable-anchor';
 
+const HEADER_STATES = {
+  HOME: 'home',
+  ABOUT: 'about',
+  MUSIC: 'music',
+  GALLERY: 'gallery',
+  CONTACT: 'contact'
+};
+
 export default class Header extends Component {
-  static renderHeaderButton(anchor, label) {
+  constructor(props) {
+    super(props);
+    this.state = { header: HEADER_STATES.HOME };
+    this.onHashChange = this.handleHashChange.bind(this);
+  }
+
+  componentDidMount() {
+    window.addEventListener('hashchange', this.onHashChange);
+  }
+
+  componentWillUnmount() {
+    window.removeEventListener('hashchange', this.onHashChange);
+  }
+
+  handleHashChange() {
+    this.setState({ header: window.location.hash.substring(1) });
+  }
+
+  renderHeaderButton(anchor, label) {
+    const { header } = this.state;
     return (
       <button
-        className="header-button"
+        className={classNames({
+          'header-button': true,
+          'header-button-selected': header === anchor
+        })}
         onClick={() => { removeHash(); goToAnchor(anchor); }}
       >
         {label}
@@ -17,11 +48,11 @@ export default class Header extends Component {
     return (
       <div className="header-container">
         <div className="header">
-          {Header.renderHeaderButton('home', 'HOME')}
-          {Header.renderHeaderButton('about', 'ABOUT')}
-          {Header.renderHeaderButton('music', 'MUSIC')}
-          {Header.renderHeaderButton('gallery', 'GALLERY')}
-          {Header.renderHeaderButton('contact', 'CONTACT')}
+          {this.renderHeaderButton('home', 'HOME')}
+          {this.renderHeaderButton('about', 'ABOUT')}
+          {this.renderHeaderButton('music', 'MUSIC')}
+          {this.renderHeaderButton('gallery', 'GALLERY')}
+          {this.renderHeaderButton('contact', 'CONTACT')}
         </div>
       </div>
     );
